remove exception from the base class
The current implementation tries to help users to implement the required
API. A stub of the `resize` member function allows skipping of
overriding when it is not necessary. But at the same time, this stub
implementation adds dependencies from exceptions that may be
inappropriate in some cases (embedded systems, etc).

It is proposed to remove the exception from the base class and if a stub
implementation is required a user will have to implement it.

Test: build
Issue: #20
Signed-off-by: Gluttton <gluttton@ukr.net>