simplify unit tests
The current tests implementation has an additional level of indirection
that is caused by Google Test parameters list limitation (by default its
length is 50). That's why it is necessary to split the whole parameters
into separate lists and duplicate a call of the same test.  As result
tests have a lot of repetitions and look complicated.

To avoid this issue it is proposed to modify the Google Test framework
and increase the length of the type parameter list.

Test: check tests,
      $ mkdir build
      $ cd build
      $ cmake -DMUNKRESCPP_DEVEL_MODE=ON ..
      $ make tests
      $ ./tests/munkrestest
Issue: #19
Signed-off-by: Gluttton <gluttton@ukr.net>