ci: fix coverage badge
The old configuration doesn't work well anymore and now it is mandatory
to explicitly set regex command to extract coverage info.

Since an arbitrary regex command can be used it is not necessary to use
`gcovr` anymore and coverage data can be extracted from the report
created by `lcov`.

Issue: #32
Signed-off-by: Gluttton <gluttton@ukr.net>